+++ src/main/java/CaffeineCoder/recipic/global/image/ImageService.java
@@ -1,11 +1,10 @@
-package CaffeineCoder.recipic.domain.recipe.application;
+package CaffeineCoder.recipic.global.image;
 
 import com.google.cloud.storage.BlobInfo;
 import com.google.cloud.storage.Storage;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
